refactor(cli): extract sensor parser owner
